gh actions: add more lanes
With this change, we reach parity with the travis lanes
we planned to add. We replace older go version with more modern ones.

TO make the lane consistent, add skips and/or move test
around where relevant.
