+++ pkg/pci/pci_test.go
@@ -10,8 +10,6 @@ import (
 	"os"
 	"testing"
 
-	"github.com/jaypipes/ghw/pkg/context"
-	"github.com/jaypipes/ghw/pkg/marshal"
 	"github.com/jaypipes/ghw/pkg/pci"
 )
 
@@ -19,6 +17,7 @@ func TestPCI(t *testing.T) {
 	if _, ok := os.LookupEnv("GHW_TESTING_SKIP_PCI"); ok {
 		t.Skip("Skipping PCI tests.")
 	}
+
 	info, err := pci.New()
 	if err != nil {
 		t.Fatalf("Expected no error creating PciInfo, but got %v", err)
@@ -57,21 +56,5 @@ func TestPCI(t *testing.T) {
 		if dev.ProgrammingInterface == nil {
 			t.Fatalf("Expected device programming interface for %s to be non-nil", dev.Address)
 		}
-	}
-}
-
-func TestPCIMarshalJSON(t *testing.T) {
-	if _, ok := os.LookupEnv("GHW_TESTING_SKIP_PCI"); ok {
-		t.Skip("Skipping PCI tests.")
-	}
-	info, err := pci.New()
-	if err != nil {
-		t.Fatalf("Expected no error creating PciInfo, but got %v", err)
-	}
-
-	dev := info.ParseDevice("0000:3c:00.0", "pci:v0000144Dd0000A804sv0000144Dsd0000A801bc01sc08i02")
-	s := marshal.SafeJSON(context.FromEnv(), dev, true)
-	if s == "" {
-		t.Fatalf("Error marshalling device: %v", dev)
 	}
 }
